Split Gui code into multiple files
