+++ czkawka_gui/src/connect_button_save.rs
@@ -6,7 +6,7 @@ use gtk::prelude::*;
 
 pub fn connect_button_save(gui_data: &GuiData) {
     let gui_data = gui_data.clone();
-    let buttons_save = gui_data.buttons_save.clone();
+    let buttons_save = gui_data.bottom_buttons.buttons_save.clone();
     let shared_duplication_state = gui_data.shared_duplication_state.clone();
     let shared_empty_folders_state = gui_data.shared_empty_folders_state.clone();
     let shared_big_files_state = gui_data.shared_big_files_state.clone();
@@ -16,7 +16,7 @@ pub fn connect_button_save(gui_data: &GuiData) {
     let shared_same_music_state = gui_data.shared_same_music_state.clone();
     let shared_zeroed_files_state = gui_data.shared_zeroed_files_state.clone();
     let shared_same_invalid_symlinks = gui_data.shared_same_invalid_symlinks.clone();
-    let notebook_main = gui_data.notebook_main.clone();
+    let notebook_main = gui_data.main_notebook.notebook_main.clone();
     buttons_save.connect_clicked(move |_| {
         let file_name;
 
@@ -72,7 +72,7 @@ pub fn connect_button_save(gui_data: &GuiData) {
 }
 fn post_save_things(file_name: &str, type_of_tab: &NotebookMainEnum, gui_data: &GuiData) {
     let entry_info = gui_data.entry_info.clone();
-    let buttons_save = gui_data.buttons_save.clone();
+    let buttons_save = gui_data.bottom_buttons.buttons_save.clone();
     let shared_buttons = gui_data.shared_buttons.clone();
 
     entry_info.set_text(format!("Saved results to file {}", file_name).as_str());
