+++ czkawka_gui/src/connect_button_search.rs
@@ -35,35 +35,35 @@ pub fn connect_button_search(
     futures_sender_invalid_symlinks: futures::channel::mpsc::Sender<invalid_symlinks::ProgressData>,
 ) {
     let entry_info = gui_data.entry_info.clone();
-    let notebook_main = gui_data.notebook_main.clone();
-    let tree_view_included_directories = gui_data.tree_view_included_directories.clone();
-    let tree_view_excluded_directories = gui_data.tree_view_excluded_directories.clone();
-    let buttons_search_clone = gui_data.buttons_search.clone();
-    let buttons_array = gui_data.buttons_array.clone();
-    let check_button_recursive = gui_data.check_button_recursive.clone();
-    let entry_excluded_items = gui_data.entry_excluded_items.clone();
-    let entry_same_music_minimal_size = gui_data.entry_same_music_minimal_size.clone();
-    let entry_allowed_extensions = gui_data.entry_allowed_extensions.clone();
-    let buttons_names = gui_data.buttons_names.clone();
-    let radio_button_duplicates_name = gui_data.radio_button_duplicates_name.clone();
-    let radio_button_duplicates_size = gui_data.radio_button_duplicates_size.clone();
-    let radio_button_duplicates_hashmb = gui_data.radio_button_duplicates_hashmb.clone();
-    let radio_button_duplicates_hash = gui_data.radio_button_duplicates_hash.clone();
-    let radio_button_similar_images_minimal = gui_data.radio_button_similar_images_minimal.clone();
-    let radio_button_similar_images_very_small = gui_data.radio_button_similar_images_very_small.clone();
-    let radio_button_similar_images_small = gui_data.radio_button_similar_images_small.clone();
-    let radio_button_similar_images_medium = gui_data.radio_button_similar_images_medium.clone();
-    let radio_button_similar_images_high = gui_data.radio_button_similar_images_high.clone();
-    let radio_button_similar_images_very_high = gui_data.radio_button_similar_images_very_high.clone();
-    let entry_duplicate_minimal_size = gui_data.entry_duplicate_minimal_size.clone();
+    let notebook_main = gui_data.main_notebook.notebook_main.clone();
+    let tree_view_included_directories = gui_data.upper_notebook.tree_view_included_directories.clone();
+    let tree_view_excluded_directories = gui_data.upper_notebook.tree_view_excluded_directories.clone();
+    let buttons_search_clone = gui_data.bottom_buttons.buttons_search.clone();
+    let buttons_array = gui_data.bottom_buttons.buttons_array.clone();
+    let check_button_recursive = gui_data.upper_notebook.check_button_recursive.clone();
+    let entry_excluded_items = gui_data.upper_notebook.entry_excluded_items.clone();
+    let entry_same_music_minimal_size = gui_data.main_notebook.entry_same_music_minimal_size.clone();
+    let entry_allowed_extensions = gui_data.upper_notebook.entry_allowed_extensions.clone();
+    let buttons_names = gui_data.bottom_buttons.buttons_names.clone();
+    let radio_button_duplicates_name = gui_data.main_notebook.radio_button_duplicates_name.clone();
+    let radio_button_duplicates_size = gui_data.main_notebook.radio_button_duplicates_size.clone();
+    let radio_button_duplicates_hashmb = gui_data.main_notebook.radio_button_duplicates_hashmb.clone();
+    let radio_button_duplicates_hash = gui_data.main_notebook.radio_button_duplicates_hash.clone();
+    let radio_button_similar_images_minimal = gui_data.main_notebook.radio_button_similar_images_minimal.clone();
+    let radio_button_similar_images_very_small = gui_data.main_notebook.radio_button_similar_images_very_small.clone();
+    let radio_button_similar_images_small = gui_data.main_notebook.radio_button_similar_images_small.clone();
+    let radio_button_similar_images_medium = gui_data.main_notebook.radio_button_similar_images_medium.clone();
+    let radio_button_similar_images_high = gui_data.main_notebook.radio_button_similar_images_high.clone();
+    let radio_button_similar_images_very_high = gui_data.main_notebook.radio_button_similar_images_very_high.clone();
+    let entry_duplicate_minimal_size = gui_data.main_notebook.entry_duplicate_minimal_size.clone();
     let stop_receiver = gui_data.stop_receiver.clone();
-    let entry_big_files_number = gui_data.entry_big_files_number.clone();
-    let entry_similar_images_minimal_size = gui_data.entry_similar_images_minimal_size.clone();
-    let check_button_music_title: gtk::CheckButton = gui_data.check_button_music_title.clone();
-    let check_button_music_artist: gtk::CheckButton = gui_data.check_button_music_artist.clone();
-    let check_button_music_album_title: gtk::CheckButton = gui_data.check_button_music_album_title.clone();
-    let check_button_music_album_artist: gtk::CheckButton = gui_data.check_button_music_album_artist.clone();
-    let check_button_music_year: gtk::CheckButton = gui_data.check_button_music_year.clone();
+    let entry_big_files_number = gui_data.main_notebook.entry_big_files_number.clone();
+    let entry_similar_images_minimal_size = gui_data.main_notebook.entry_similar_images_minimal_size.clone();
+    let check_button_music_title: gtk::CheckButton = gui_data.main_notebook.check_button_music_title.clone();
+    let check_button_music_artist: gtk::CheckButton = gui_data.main_notebook.check_button_music_artist.clone();
+    let check_button_music_album_title: gtk::CheckButton = gui_data.main_notebook.check_button_music_album_title.clone();
+    let check_button_music_album_artist: gtk::CheckButton = gui_data.main_notebook.check_button_music_album_artist.clone();
+    let check_button_music_year: gtk::CheckButton = gui_data.main_notebook.check_button_music_year.clone();
     let shared_buttons = gui_data.shared_buttons.clone();
     let tree_view_empty_folder_finder = gui_data.main_notebook.tree_view_empty_folder_finder.clone();
     let tree_view_empty_files_finder = gui_data.main_notebook.tree_view_empty_files_finder.clone();
@@ -75,12 +75,12 @@ pub fn connect_button_search(
     let tree_view_zeroed_files_finder = gui_data.main_notebook.tree_view_zeroed_files_finder.clone();
     let tree_view_invalid_symlinks = gui_data.main_notebook.tree_view_invalid_symlinks.clone();
     let text_view_errors = gui_data.text_view_errors.clone();
-    let dialog_progress = gui_data.dialog_progress.clone();
-    let label_stage = gui_data.label_stage.clone();
-    let grid_progress_stages = gui_data.grid_progress_stages.clone();
-    let progress_bar_current_stage = gui_data.progress_bar_current_stage.clone();
-    let progress_bar_all_stages = gui_data.progress_bar_all_stages.clone();
-    let image_preview_similar_images = gui_data.image_preview_similar_images.clone();
+    let dialog_progress = gui_data.progress_dialog.dialog_progress.clone();
+    let label_stage = gui_data.progress_dialog.label_stage.clone();
+    let grid_progress_stages = gui_data.progress_dialog.grid_progress_stages.clone();
+    let progress_bar_current_stage = gui_data.progress_dialog.progress_bar_current_stage.clone();
+    let progress_bar_all_stages = gui_data.progress_dialog.progress_bar_all_stages.clone();
+    let image_preview_similar_images = gui_data.main_notebook.image_preview_similar_images.clone();
 
     buttons_search_clone.connect_clicked(move |_| {
         let included_directories = get_string_from_list_store(&tree_view_included_directories);
