+++ czkawka_gui/src/connect_hide_text_view_errors.rs
@@ -3,8 +3,8 @@ use crate::gui_data::GuiData;
 use gtk::prelude::*;
 
 pub fn connect_hide_text_view_errors(gui_data: &GuiData) {
-    let check_button_settings_show_text_view = gui_data.check_button_settings_show_text_view.clone();
-    let buttons_show_errors = gui_data.buttons_show_errors.clone();
+    let check_button_settings_show_text_view = gui_data.upper_notebook.check_button_settings_show_text_view.clone();
+    let buttons_show_errors = gui_data.bottom_buttons.buttons_show_errors.clone();
     let scrolled_window_errors = gui_data.scrolled_window_errors.clone();
 
     buttons_show_errors.connect_clicked(move |_| {
