+++ czkawka_gui/src/connect_notebook_tabs.rs
@@ -6,11 +6,11 @@ use gtk::prelude::*;
 
 pub fn connect_notebook_tabs(gui_data: &GuiData) {
     let shared_buttons = gui_data.shared_buttons.clone();
-    let buttons_array = gui_data.buttons_array.clone();
-    let notebook_main_clone = gui_data.notebook_main.clone();
-    let buttons_names = gui_data.buttons_names.clone();
+    let buttons_array = gui_data.bottom_buttons.buttons_array.clone();
+    let notebook_main_clone = gui_data.main_notebook.notebook_main.clone();
+    let buttons_names = gui_data.bottom_buttons.buttons_names.clone();
     let shared_upper_notebooks = gui_data.shared_upper_notebooks.clone();
-    let notebook_upper = gui_data.notebook_upper.clone();
+    let notebook_upper = gui_data.upper_notebook.notebook_upper.clone();
 
     notebook_main_clone.connect_switch_page(move |_, _, number| {
         let current_tab_in_main_notebook = to_notebook_main_enum(number);
