+++ czkawka_gui/src/connect_progress_window.rs
@@ -22,10 +22,10 @@ pub fn connect_progress_window(
 
     {
         // Duplicate Files
-        let label_stage = gui_data.label_stage.clone();
-        let progress_bar_current_stage = gui_data.progress_bar_current_stage.clone();
-        let progress_bar_all_stages = gui_data.progress_bar_all_stages.clone();
-        let grid_progress_stages = gui_data.grid_progress_stages.clone();
+        let label_stage = gui_data.progress_dialog.label_stage.clone();
+        let progress_bar_current_stage = gui_data.progress_dialog.progress_bar_current_stage.clone();
+        let progress_bar_all_stages = gui_data.progress_dialog.progress_bar_all_stages.clone();
+        let grid_progress_stages = gui_data.progress_dialog.grid_progress_stages.clone();
         let future = async move {
             while let Some(item) = futures_receiver_duplicate_files.next().await {
                 match item.checking_method {
@@ -90,7 +90,7 @@ pub fn connect_progress_window(
     }
     {
         // Empty Files
-        let label_stage = gui_data.label_stage.clone();
+        let label_stage = gui_data.progress_dialog.label_stage.clone();
         let future = async move {
             while let Some(item) = futures_receiver_empty_files.next().await {
                 label_stage.set_text(format!("Scanned {} files", item.files_checked).as_str());
@@ -100,7 +100,7 @@ pub fn connect_progress_window(
     }
     {
         // Empty Folder
-        let label_stage = gui_data.label_stage.clone();
+        let label_stage = gui_data.progress_dialog.label_stage.clone();
         let future = async move {
             while let Some(item) = futures_receiver_empty_folder.next().await {
                 label_stage.set_text(format!("Scanned {} folders", item.folders_checked).as_str());
@@ -110,7 +110,7 @@ pub fn connect_progress_window(
     }
     {
         // Big Files
-        let label_stage = gui_data.label_stage.clone();
+        let label_stage = gui_data.progress_dialog.label_stage.clone();
         let future = async move {
             while let Some(item) = futures_receiver_big_files.next().await {
                 label_stage.set_text(format!("Scanned {} files", item.files_checked).as_str());
@@ -120,9 +120,9 @@ pub fn connect_progress_window(
     }
     {
         // Same Music
-        let label_stage = gui_data.label_stage.clone();
-        let progress_bar_current_stage = gui_data.progress_bar_current_stage.clone();
-        let progress_bar_all_stages = gui_data.progress_bar_all_stages.clone();
+        let label_stage = gui_data.progress_dialog.label_stage.clone();
+        let progress_bar_current_stage = gui_data.progress_dialog.progress_bar_current_stage.clone();
+        let progress_bar_all_stages = gui_data.progress_dialog.progress_bar_all_stages.clone();
         let future = async move {
             while let Some(item) = futures_receiver_same_music.next().await {
                 match item.current_stage {
@@ -161,9 +161,9 @@ pub fn connect_progress_window(
     }
     {
         // Similar Images
-        let label_stage = gui_data.label_stage.clone();
-        let progress_bar_current_stage = gui_data.progress_bar_current_stage.clone();
-        let progress_bar_all_stages = gui_data.progress_bar_all_stages.clone();
+        let label_stage = gui_data.progress_dialog.label_stage.clone();
+        let progress_bar_current_stage = gui_data.progress_dialog.progress_bar_current_stage.clone();
+        let progress_bar_all_stages = gui_data.progress_dialog.progress_bar_all_stages.clone();
         let future = async move {
             while let Some(item) = futures_receiver_similar_images.next().await {
                 match item.current_stage {
@@ -192,7 +192,7 @@ pub fn connect_progress_window(
     }
     {
         // Temporary
-        let label_stage = gui_data.label_stage.clone();
+        let label_stage = gui_data.progress_dialog.label_stage.clone();
         let future = async move {
             while let Some(item) = futures_receiver_temporary.next().await {
                 label_stage.set_text(format!("Scanned {} files", item.files_checked).as_str());
@@ -202,9 +202,9 @@ pub fn connect_progress_window(
     }
     {
         // Zeroed Files
-        let label_stage = gui_data.label_stage.clone();
-        let progress_bar_current_stage = gui_data.progress_bar_current_stage.clone();
-        let progress_bar_all_stages = gui_data.progress_bar_all_stages.clone();
+        let label_stage = gui_data.progress_dialog.label_stage.clone();
+        let progress_bar_current_stage = gui_data.progress_dialog.progress_bar_current_stage.clone();
+        let progress_bar_all_stages = gui_data.progress_dialog.progress_bar_all_stages.clone();
         let future = async move {
             while let Some(item) = futures_receiver_zeroed.next().await {
                 match item.current_stage {
@@ -233,7 +233,7 @@ pub fn connect_progress_window(
     }
     {
         // Invalid Symlinks
-        let label_stage = gui_data.label_stage.clone();
+        let label_stage = gui_data.progress_dialog.label_stage.clone();
         let future = async move {
             while let Some(item) = futures_receiver_invalid_symlinks.next().await {
                 label_stage.set_text(format!("Scanned {} files", item.files_checked).as_str());
