+++ czkawka_gui/src/connect_settings.rs
@@ -7,7 +7,7 @@ pub fn connect_settings(gui_data: &GuiData) {
     // Connect save configuration button
     {
         let gui_data = gui_data.clone();
-        let button_settings_save_configuration = gui_data.button_settings_save_configuration.clone();
+        let button_settings_save_configuration = gui_data.upper_notebook.button_settings_save_configuration.clone();
         button_settings_save_configuration.connect_clicked(move |_| {
             save_configuration(&gui_data, true);
         });
@@ -15,7 +15,7 @@ pub fn connect_settings(gui_data: &GuiData) {
     // Connect load configuration button
     {
         let gui_data = gui_data.clone();
-        let button_settings_load_configuration = gui_data.button_settings_load_configuration.clone();
+        let button_settings_load_configuration = gui_data.upper_notebook.button_settings_load_configuration.clone();
         button_settings_load_configuration.connect_clicked(move |_| {
             load_configuration(&gui_data, true);
         });
@@ -23,7 +23,7 @@ pub fn connect_settings(gui_data: &GuiData) {
     // Connect reset configuration button
     {
         let gui_data = gui_data.clone();
-        let button_settings_reset_configuration = gui_data.button_settings_reset_configuration.clone();
+        let button_settings_reset_configuration = gui_data.upper_notebook.button_settings_reset_configuration.clone();
         button_settings_reset_configuration.connect_clicked(move |_| {
             reset_configuration(&gui_data, true);
         });
