+++ czkawka_gui/src/gui_data.rs
@@ -1,5 +1,9 @@
 extern crate gtk;
+use crate::gui_bottom_buttons::GUIBottomButtons;
 use crate::gui_main_notebook::GUIMainNotebook;
+use crate::gui_popovers::GUIPopovers;
+use crate::gui_progress_dialog::GUIProgressDialog;
+use crate::gui_upper_notepad::GUIUpperNotebook;
 use crate::notebook_enums::*;
 use crossbeam_channel::unbounded;
 use czkawka_core::big_file::BigFile;
@@ -12,7 +16,7 @@ use czkawka_core::similar_images::SimilarImages;
 use czkawka_core::temporary::Temporary;
 use czkawka_core::zeroed::ZeroedFiles;
 use gtk::prelude::*;
-use gtk::{Builder, Button, TreeView};
+use gtk::Builder;
 use std::cell::RefCell;
 use std::collections::HashMap;
 use std::rc::Rc;
@@ -27,9 +31,10 @@ pub struct GuiData {
     pub window_main: gtk::Window,
 
     pub main_notebook: GUIMainNotebook,
-
-    // States
-    pub buttons_labels: [String; 5],
+    pub upper_notebook: GUIUpperNotebook,
+    pub popovers: GUIPopovers,
+    pub bottom_buttons: GUIBottomButtons,
+    pub progress_dialog: GUIProgressDialog,
 
     // Buttons state
     pub shared_buttons: Rc<RefCell<HashMap<NotebookMainEnum, HashMap<String, bool>>>>,
@@ -48,127 +53,12 @@ pub struct GuiData {
     pub shared_same_music_state: Rc<RefCell<SameMusic>>,
     pub shared_same_invalid_symlinks: Rc<RefCell<InvalidSymlinks>>,
 
-    //// GUI Entry
-    pub entry_similar_images_minimal_size: gtk::Entry,
-    pub entry_duplicate_minimal_size: gtk::Entry,
-    pub entry_allowed_extensions: gtk::Entry,
-    pub entry_excluded_items: gtk::Entry,
-    pub entry_big_files_number: gtk::Entry,
-    pub entry_same_music_minimal_size: gtk::Entry,
-
-    //// GUI Buttons
-    pub buttons_search: gtk::Button,
-    pub buttons_select: gtk::Button,
-    pub buttons_delete: gtk::Button,
-    pub buttons_save: gtk::Button,
-    pub buttons_symlink: gtk::Button,
-    pub buttons_show_errors: gtk::Button,
-    pub buttons_names: [String; 5],
-    pub buttons_array: [Button; 5],
-
-    pub buttons_manual_add_directory: gtk::Button,
-    pub buttons_add_included_directory: gtk::Button,
-    pub buttons_remove_included_directory: gtk::Button,
-    pub buttons_manual_add_excluded_directory: gtk::Button,
-    pub buttons_add_excluded_directory: gtk::Button,
-    pub buttons_remove_excluded_directory: gtk::Button,
-
-    // Buttons search popover buttons
-    pub buttons_popover_select_all: gtk::Button,
-    pub buttons_popover_unselect_all: gtk::Button,
-    pub buttons_popover_reverse: gtk::Button,
-    pub buttons_popover_select_all_except_oldest: gtk::Button,
-    pub buttons_popover_select_all_except_newest: gtk::Button,
-    pub buttons_popover_select_one_oldest: gtk::Button,
-    pub buttons_popover_select_one_newest: gtk::Button,
-    pub buttons_popover_select_custom: gtk::Button,
-    pub buttons_popover_unselect_custom: gtk::Button,
-    pub buttons_popover_select_all_images_except_biggest: gtk::Button,
-    pub buttons_popover_select_all_images_except_smallest: gtk::Button,
-
-    pub separator_select_image_size: gtk::Separator,
-    pub separator_select_reverse: gtk::Separator,
-    pub separator_select_date: gtk::Separator,
-    pub separator_select_custom: gtk::Separator,
-
-    pub buttons_popover_right_click_open_file: gtk::Button,
-    pub buttons_popover_right_click_open_folder: gtk::Button,
-
-    //// Popovers
-    pub popover_select: gtk::Popover,
-    pub popover_right_click: gtk::Popover,
-
-    //// Check Buttons
-    pub check_button_recursive: gtk::CheckButton,
-
-    pub check_button_music_title: gtk::CheckButton,
-    pub check_button_music_artist: gtk::CheckButton,
-    pub check_button_music_album_title: gtk::CheckButton,
-    pub check_button_music_album_artist: gtk::CheckButton,
-    pub check_button_music_year: gtk::CheckButton,
-
-    //// Radio Buttons
-    // Duplicates
-    pub radio_button_duplicates_name: gtk::RadioButton,
-    pub radio_button_duplicates_size: gtk::RadioButton,
-    pub radio_button_duplicates_hashmb: gtk::RadioButton,
-    pub radio_button_duplicates_hash: gtk::RadioButton,
-
-    pub radio_button_similar_images_minimal: gtk::RadioButton,
-    pub radio_button_similar_images_very_small: gtk::RadioButton,
-    pub radio_button_similar_images_small: gtk::RadioButton,
-    pub radio_button_similar_images_medium: gtk::RadioButton,
-    pub radio_button_similar_images_high: gtk::RadioButton,
-    pub radio_button_similar_images_very_high: gtk::RadioButton,
-
-    //// Notebooks
-    pub notebook_main: gtk::Notebook,
-    pub notebook_upper: gtk::Notebook,
-
-    pub notebook_upper_children_names: Vec<String>,
-
     //// Entry
-    pub entry_info: gtk::Entry, // To show default
+    pub entry_info: gtk::Entry,
 
     //// Bottom
     pub text_view_errors: gtk::TextView,
     pub scrolled_window_errors: gtk::ScrolledWindow,
-
-    //// Scrolled windows
-
-    // Upper notebook
-    pub scrolled_window_included_directories: gtk::ScrolledWindow,
-    pub scrolled_window_excluded_directories: gtk::ScrolledWindow,
-
-    pub tree_view_included_directories: gtk::TreeView,
-    pub tree_view_excluded_directories: gtk::TreeView,
-
-    //// Dialog State - dialog with progress state, which allows to stop task
-    pub dialog_progress: gtk::Dialog,
-
-    pub progress_bar_current_stage: gtk::ProgressBar,
-    pub progress_bar_all_stages: gtk::ProgressBar,
-
-    pub label_stage: gtk::Label,
-
-    pub grid_progress_stages: gtk::Grid,
-
-    pub button_stop_in_dialog: gtk::Button,
-
-    //// Similar Images
-    pub image_preview_similar_images: gtk::Image,
-
-    //// Settings
-    pub check_button_settings_save_at_exit: gtk::CheckButton,
-    pub check_button_settings_load_at_start: gtk::CheckButton,
-    pub check_button_settings_confirm_deletion: gtk::CheckButton,
-    pub check_button_settings_show_preview_similar_images: gtk::CheckButton,
-    pub check_button_settings_show_text_view: gtk::CheckButton,
-
-    pub button_settings_save_configuration: gtk::Button,
-    pub button_settings_load_configuration: gtk::Button,
-    pub button_settings_reset_configuration: gtk::Button,
-    //// Threads
 
     // Used for sending stop signal to thread
     pub stop_sender: crossbeam_channel::Sender<()>,
@@ -187,23 +77,24 @@ impl GuiData {
         window_main.set_title("Czkawka");
 
         let main_notebook = GUIMainNotebook::create_from_builder(&builder);
+        let upper_notebook = GUIUpperNotebook::create_from_builder(&builder);
+        let popovers = GUIPopovers::create_from_builder(&builder);
+        let bottom_buttons = GUIBottomButtons::create_from_builder(&builder);
+        let progress_dialog = GUIProgressDialog::create_from_builder(&builder);
 
         ////////////////////////////////////////////////////////////////////////////////////////////////
-        //// States
-        let buttons_labels = ["search".to_string(), "select".to_string(), "delete".to_string(), "save".to_string(), "symlink".to_string()];
 
         // Buttons State - to remember existence of different buttons on pages
-
         let shared_buttons: Rc<RefCell<_>> = Rc::new(RefCell::new(HashMap::<NotebookMainEnum, HashMap<String, bool>>::new()));
 
         // Show by default only search button
         for i in get_all_main_tabs().iter() {
             let mut temp_hashmap: HashMap<String, bool> = Default::default();
-            for j in buttons_labels.iter() {
-                if *j == "search" {
-                    temp_hashmap.insert(j.to_string(), true);
+            for button_name in bottom_buttons.buttons_names.iter() {
+                if *button_name == "search" {
+                    temp_hashmap.insert(button_name.to_string(), true);
                 } else {
-                    temp_hashmap.insert(j.to_string(), false);
+                    temp_hashmap.insert(button_name.to_string(), false);
                 }
             }
             shared_buttons.borrow_mut().insert(i.clone(), temp_hashmap);
@@ -234,137 +125,12 @@ impl GuiData {
         let shared_same_music_state: Rc<RefCell<_>> = Rc::new(RefCell::new(SameMusic::new()));
         let shared_same_invalid_symlinks: Rc<RefCell<_>> = Rc::new(RefCell::new(InvalidSymlinks::new()));
 
-        ////////////////////////////////////////////////////////////////////////////////////////////////
-
-        //// GUI Entry
-        let entry_similar_images_minimal_size: gtk::Entry = builder.get_object("entry_similar_images_minimal_size").unwrap();
-        let entry_duplicate_minimal_size: gtk::Entry = builder.get_object("entry_duplicate_minimal_size").unwrap();
-        let entry_allowed_extensions: gtk::Entry = builder.get_object("entry_allowed_extensions").unwrap();
-        let entry_excluded_items: gtk::Entry = builder.get_object("entry_excluded_items").unwrap();
-        let entry_big_files_number: gtk::Entry = builder.get_object("entry_big_files_number").unwrap();
-        let entry_same_music_minimal_size: gtk::Entry = builder.get_object("entry_same_music_minimal_size").unwrap();
-
-        //// GUI Buttons
-        let buttons_search: gtk::Button = builder.get_object("buttons_search").unwrap();
-        let buttons_select: gtk::Button = builder.get_object("buttons_select").unwrap();
-        let buttons_delete: gtk::Button = builder.get_object("buttons_delete").unwrap();
-        let buttons_save: gtk::Button = builder.get_object("buttons_save").unwrap();
-        let buttons_symlink: gtk::Button = builder.get_object("buttons_symlink").unwrap();
-
-        let buttons_show_errors: gtk::Button = builder.get_object("buttons_show_errors").unwrap();
-
-        let buttons_names = ["search".to_string(), "select".to_string(), "delete".to_string(), "save".to_string(), "symlink".to_string()];
-        let buttons_array = [buttons_search.clone(), buttons_select.clone(), buttons_delete.clone(), buttons_save.clone(), buttons_symlink.clone()];
-
-        let buttons_manual_add_directory: gtk::Button = builder.get_object("buttons_manual_add_directory").unwrap();
-        let buttons_add_included_directory: gtk::Button = builder.get_object("buttons_add_included_directory").unwrap();
-        let buttons_remove_included_directory: gtk::Button = builder.get_object("buttons_remove_included_directory").unwrap();
-        let buttons_manual_add_excluded_directory: gtk::Button = builder.get_object("buttons_manual_add_excluded_directory").unwrap();
-        let buttons_add_excluded_directory: gtk::Button = builder.get_object("buttons_add_excluded_directory").unwrap();
-        let buttons_remove_excluded_directory: gtk::Button = builder.get_object("buttons_remove_excluded_directory").unwrap();
-
-        // Buttons search popover buttons
-        let buttons_popover_select_all: gtk::Button = builder.get_object("buttons_popover_select_all").unwrap();
-        let buttons_popover_unselect_all: gtk::Button = builder.get_object("buttons_popover_unselect_all").unwrap();
-        let buttons_popover_reverse: gtk::Button = builder.get_object("buttons_popover_reverse").unwrap();
-        let buttons_popover_select_all_except_oldest: gtk::Button = builder.get_object("buttons_popover_select_all_except_oldest").unwrap();
-        let buttons_popover_select_all_except_newest: gtk::Button = builder.get_object("buttons_popover_select_all_except_newest").unwrap();
-        let buttons_popover_select_one_oldest: gtk::Button = builder.get_object("buttons_popover_select_one_oldest").unwrap();
-        let buttons_popover_select_one_newest: gtk::Button = builder.get_object("buttons_popover_select_one_newest").unwrap();
-        let buttons_popover_select_custom: gtk::Button = builder.get_object("buttons_popover_select_custom").unwrap();
-        let buttons_popover_unselect_custom: gtk::Button = builder.get_object("buttons_popover_unselect_custom").unwrap();
-        let buttons_popover_select_all_images_except_biggest: gtk::Button = builder.get_object("buttons_popover_select_all_images_except_biggest").unwrap();
-        let buttons_popover_select_all_images_except_smallest: gtk::Button = builder.get_object("buttons_popover_select_all_images_except_smallest").unwrap();
-
-        let separator_select_image_size: gtk::Separator = builder.get_object("separator_select_image_size").unwrap();
-        let separator_select_reverse: gtk::Separator = builder.get_object("separator_select_reverse").unwrap();
-        let separator_select_date: gtk::Separator = builder.get_object("separator_select_date").unwrap();
-        let separator_select_custom: gtk::Separator = builder.get_object("separator_select_custom").unwrap();
-
-        let buttons_popover_right_click_open_file: gtk::Button = builder.get_object("buttons_popover_right_click_open_file").unwrap();
-        let buttons_popover_right_click_open_folder: gtk::Button = builder.get_object("buttons_popover_right_click_open_folder").unwrap();
-
-        //// Popovers
-        let popover_select: gtk::Popover = builder.get_object("popover_select").unwrap();
-        let popover_right_click: gtk::Popover = builder.get_object("popover_right_click").unwrap();
-
-        //// Check Buttons
-        let check_button_recursive: gtk::CheckButton = builder.get_object("check_button_recursive").unwrap();
-        let check_button_music_title: gtk::CheckButton = builder.get_object("check_button_music_title").unwrap();
-        let check_button_music_artist: gtk::CheckButton = builder.get_object("check_button_music_artist").unwrap();
-        let check_button_music_album_title: gtk::CheckButton = builder.get_object("check_button_music_album_title").unwrap();
-        let check_button_music_album_artist: gtk::CheckButton = builder.get_object("check_button_music_album_artist").unwrap();
-        let check_button_music_year: gtk::CheckButton = builder.get_object("check_button_music_year").unwrap();
-
-        //// Radio Buttons
-        let radio_button_duplicates_name: gtk::RadioButton = builder.get_object("radio_button_duplicates_name").unwrap();
-        let radio_button_duplicates_size: gtk::RadioButton = builder.get_object("radio_button_duplicates_size").unwrap();
-        let radio_button_duplicates_hashmb: gtk::RadioButton = builder.get_object("radio_button_duplicates_hashmb").unwrap();
-        let radio_button_duplicates_hash: gtk::RadioButton = builder.get_object("radio_button_duplicates_hash").unwrap();
-
-        let radio_button_similar_images_minimal: gtk::RadioButton = builder.get_object("radio_button_similar_images_minimal").unwrap();
-        let radio_button_similar_images_very_small: gtk::RadioButton = builder.get_object("radio_button_similar_images_very_small").unwrap();
-        let radio_button_similar_images_small: gtk::RadioButton = builder.get_object("radio_button_similar_images_small").unwrap();
-        let radio_button_similar_images_medium: gtk::RadioButton = builder.get_object("radio_button_similar_images_medium").unwrap();
-        let radio_button_similar_images_high: gtk::RadioButton = builder.get_object("radio_button_similar_images_high").unwrap();
-        let radio_button_similar_images_very_high: gtk::RadioButton = builder.get_object("radio_button_similar_images_very_high").unwrap();
-
-        //// Notebooks
-        let notebook_main: gtk::Notebook = builder.get_object("notebook_main").unwrap();
-        let notebook_upper: gtk::Notebook = builder.get_object("notebook_upper").unwrap();
-
-        let mut notebook_upper_children_names: Vec<String> = Vec::new();
-
-        for i in notebook_upper.get_children() {
-            notebook_upper_children_names.push(i.get_buildable_name().unwrap().to_string());
-        }
-
         //// Entry
         let entry_info: gtk::Entry = builder.get_object("entry_info").unwrap();
 
         //// Bottom
         let text_view_errors: gtk::TextView = builder.get_object("text_view_errors").unwrap();
         let scrolled_window_errors: gtk::ScrolledWindow = builder.get_object("scrolled_window_errors").unwrap();
-
-        //// Scrolled windows
-        // Main notebook
-
-        // Upper notebook
-        let scrolled_window_included_directories: gtk::ScrolledWindow = builder.get_object("scrolled_window_included_directories").unwrap();
-        let scrolled_window_excluded_directories: gtk::ScrolledWindow = builder.get_object("scrolled_window_excluded_directories").unwrap();
-
-        let tree_view_included_directories: gtk::TreeView = TreeView::new();
-        let tree_view_excluded_directories: gtk::TreeView = TreeView::new();
-
-        //// Dialog State - dialog with progress state, which allows to stop task
-        let dialog_progress: gtk::Dialog = builder.get_object("dialog_progress").unwrap();
-        dialog_progress.set_title("Czkawka");
-
-        let progress_bar_current_stage: gtk::ProgressBar = builder.get_object("progress_bar_current_stage").unwrap();
-        let progress_bar_all_stages: gtk::ProgressBar = builder.get_object("progress_bar_all_stages").unwrap();
-
-        let label_stage: gtk::Label = builder.get_object("label_stage").unwrap();
-
-        let grid_progress_stages: gtk::Grid = builder.get_object("grid_progress_stages").unwrap();
-
-        let button_stop_in_dialog: gtk::Button = builder.get_object("button_stop_in_dialog").unwrap();
-
-        //// Similar Images
-        let image_preview_similar_images: gtk::Image = builder.get_object("image_preview_similar_images").unwrap();
-
-        //// Settings
-        let check_button_settings_save_at_exit: gtk::CheckButton = builder.get_object("check_button_settings_save_at_exit").unwrap();
-        let check_button_settings_load_at_start: gtk::CheckButton = builder.get_object("check_button_settings_load_at_start").unwrap();
-        let check_button_settings_confirm_deletion: gtk::CheckButton = builder.get_object("check_button_settings_confirm_deletion").unwrap();
-        let check_button_settings_show_preview_similar_images: gtk::CheckButton = builder.get_object("check_button_settings_show_preview_similar_images").unwrap();
-        let check_button_settings_show_text_view: gtk::CheckButton = builder.get_object("check_button_settings_show_text_view").unwrap();
-
-        let button_settings_save_configuration: gtk::Button = builder.get_object("button_settings_save_configuration").unwrap();
-        let button_settings_load_configuration: gtk::Button = builder.get_object("button_settings_load_configuration").unwrap();
-        let button_settings_reset_configuration: gtk::Button = builder.get_object("button_settings_reset_configuration").unwrap();
-
-        //// Threads
-        // Types of messages to send to main thread where gui can be draw.
 
         // Used for sending stop signal to thread
         let (stop_sender, stop_receiver): (crossbeam_channel::Sender<()>, crossbeam_channel::Receiver<()>) = unbounded();
@@ -374,7 +140,10 @@ impl GuiData {
             builder,
             window_main,
             main_notebook,
-            buttons_labels,
+            upper_notebook,
+            popovers,
+            bottom_buttons,
+            progress_dialog,
             shared_buttons,
             shared_upper_notebooks,
             shared_duplication_state,
@@ -386,86 +155,9 @@ impl GuiData {
             shared_zeroed_files_state,
             shared_same_music_state,
             shared_same_invalid_symlinks,
-            entry_similar_images_minimal_size,
-            entry_duplicate_minimal_size,
-            entry_allowed_extensions,
-            entry_excluded_items,
-            entry_big_files_number,
-            entry_same_music_minimal_size,
-            buttons_search,
-            buttons_select,
-            buttons_delete,
-            buttons_save,
-            buttons_symlink,
-            buttons_show_errors,
-            buttons_names,
-            buttons_array,
-            buttons_manual_add_directory,
-            buttons_add_included_directory,
-            buttons_remove_included_directory,
-            buttons_manual_add_excluded_directory,
-            buttons_add_excluded_directory,
-            buttons_remove_excluded_directory,
-            buttons_popover_select_all,
-            buttons_popover_unselect_all,
-            buttons_popover_reverse,
-            buttons_popover_select_all_except_oldest,
-            buttons_popover_select_all_except_newest,
-            buttons_popover_select_one_oldest,
-            buttons_popover_select_one_newest,
-            buttons_popover_select_custom,
-            buttons_popover_unselect_custom,
-            buttons_popover_select_all_images_except_biggest,
-            buttons_popover_select_all_images_except_smallest,
-            separator_select_image_size,
-            separator_select_reverse,
-            separator_select_date,
-            separator_select_custom,
-            buttons_popover_right_click_open_file,
-            buttons_popover_right_click_open_folder,
-            popover_select,
-            popover_right_click,
-            check_button_recursive,
-            check_button_music_title,
-            check_button_music_artist,
-            check_button_music_album_title,
-            check_button_music_album_artist,
-            check_button_music_year,
-            radio_button_duplicates_name,
-            radio_button_duplicates_size,
-            radio_button_duplicates_hashmb,
-            radio_button_duplicates_hash,
-            radio_button_similar_images_minimal,
-            radio_button_similar_images_very_small,
-            radio_button_similar_images_small,
-            radio_button_similar_images_medium,
-            radio_button_similar_images_high,
-            radio_button_similar_images_very_high,
-            notebook_main,
-            notebook_upper,
-            notebook_upper_children_names,
             entry_info,
             text_view_errors,
             scrolled_window_errors,
-            scrolled_window_included_directories,
-            scrolled_window_excluded_directories,
-            tree_view_included_directories,
-            tree_view_excluded_directories,
-            dialog_progress,
-            progress_bar_current_stage,
-            progress_bar_all_stages,
-            label_stage,
-            grid_progress_stages,
-            button_stop_in_dialog,
-            image_preview_similar_images,
-            check_button_settings_save_at_exit,
-            check_button_settings_load_at_start,
-            check_button_settings_confirm_deletion,
-            check_button_settings_show_preview_similar_images,
-            check_button_settings_show_text_view,
-            button_settings_save_configuration,
-            button_settings_load_configuration,
-            button_settings_reset_configuration,
             stop_sender,
             stop_receiver,
         }
