+++ czkawka_gui/src/main.rs
@@ -16,8 +16,12 @@ mod connect_selection_of_directories;
 mod connect_settings;
 mod create_tree_view;
 mod double_click_opening;
+mod gui_bottom_buttons;
 mod gui_data;
 mod gui_main_notebook;
+mod gui_popovers;
+mod gui_progress_dialog;
+mod gui_upper_notepad;
 mod help_functions;
 mod initialize_gui;
 mod notebook_enums;
